pyproject.toml
[project]
name = "test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiokafka>=0.14.0",
    "asyncpg>=0.31.0",
    "langchain>=1.2.17",
    "langgraph>=1.1.10",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "testcontainers[kafka,postgres]>=4.14.2",
]
